PFA-23 Reorganize components folder so they are grouped by feature
